refactor: Main 컴포넌트 분리 및 상태 관리 컴포넌트 생성
- TodoContainer 컴포넌트로 분리
